Added .editorconfig to be used for Fabric projects, and updated the IDEA formatter. (#4)
Mostly adds a basic .editorconfig file to be used by people that aren't using IDEA and therefore the IDEA formatter. Obviously has way fewer features, but it is supported by every decent editor that exists. Once this gets merged, I will PR this file to the other repos as well. This should cover all files that we care about.

Also updates the IDEA formatter to stuff in the latest version.